// src/clients/api.models.ts
import type { APIRequestContext, APIResponse } from "@playwright/test";


export interface SignupPayload {
  [key: string]: string;
  name: string;
  email: string;
  password: string;
  title: string;
  birth_date: string;
  birth_month: string;
  birth_year: string;
  firstname: string;
  lastname: string;
  company: string;
  address1: string;
  address2: string;
  country: string;
  zipcode: string;
  state: string;
  city: string;
  mobile_number: string;
}


export type LoginPayload = {
  email: string;
  password: string;
};

export type LoginResponse = {
  token?: string;
  message?: string;
};